Specify scope while update progress. (#6848)
